+++ pyrate_limiter/request_rate.py
@@ -13,10 +13,17 @@ class ResetTypes(Enum):
 
 
 class RequestRate:
+    """Request rate definition.
+
+    Args:
+        limit: Number of requests allowed within ``interval``
+        interval: Time interval, in seconds
+    """
+
     def __init__(
         self,
-        limit,
-        interval,
+        limit: int,
+        interval: int,
         reset: ResetTypes = ResetTypes.INTERVAL,
     ):
         self._limit = limit
@@ -25,7 +32,7 @@ class RequestRate:
         self._log: Dict[Any, Any] = {}
 
     @property
-    def limit(self):
+    def limit(self) -> int:
         return self._limit
 
     @limit.setter
@@ -33,7 +40,7 @@ class RequestRate:
         raise ImmutableClassProperty(self, "limit")
 
     @property
-    def interval(self):
+    def interval(self) -> int:
         return self._interval
 
     @interval.setter